pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fuseiot"
version = "1.0.0"
description = "Deterministic device control SDK for Python"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "FuseIoT Team", email = "team@fuseiot.dev"}
]
keywords = ["iot", "devices", "automation", "control", "hardware", "mqtt", "http"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
]

dependencies = [
    "structlog>=22.0.0",
    "pydantic>=1.10.0",
    "PyYAML>=6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
http = ["aiohttp>=3.8.0", "requests>=2.28.0"]
mqtt = ["paho-mqtt>=1.6.0"]
websocket = ["websockets>=10.0"]
serial = ["pyserial>=3.5"]
redis = ["redis>=4.0.0"]
discovery = ["zeroconf>=0.47.0"]
monitoring = ["prometheus-client>=0.15.0"]
cloud-aws = ["boto3>=1.26.0", "AWSIoTPythonSDK>=1.5.0"]
cloud-azure = ["azure-iot-device>=2.12.0"]
cloud-gcp = ["google-cloud-iot>=2.6.0"]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
]
dev = [
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "pytest-cov>=4.0.0",
]
all = [
    "aiohttp>=3.8.0",
    "requests>=2.28.0",
    "paho-mqtt>=1.6.0",
    "websockets>=10.0",
    "pyserial>=3.5",
    "redis>=4.0.0",
    "zeroconf>=0.47.0",
    "prometheus-client>=0.15.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
]

[project.scripts]
fuseiot = "fuseiot.cli.main:main"

[project.urls]
Homepage = "https://github.com/fuseiot/fuseiot"
Documentation = "https://fuseiot.readthedocs.io"
Repository = "https://github.com/fuseiot/fuseiot.git"
Issues = "https://github.com/fuseiot/fuseiot/issues"
Changelog = "https://github.com/fuseiot/fuseiot/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["fuseiot"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers"
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (may be slow)",
    "slow: marks tests as slow",
]